refactor(achievements): centralize join logic and stabilize test environment
